Move wrapper and nav components
