+++ src/components/theme-legacy/footer.js
