+++ src/routes.js
@@ -12,7 +12,7 @@ import SearchPage from './pages/search'
 import PetitionCreatorDashboard from './containers/petition-creator-dashboard'
 import CreatePetitionPage from './pages/create-petition-page'
 import PetitionReportPage from './pages/petition-report-page'
-import Wrapper from './components/wrapper'
+import Wrapper from 'LegacyTheme/wrapper'
 import Register from './pages/register'
 import ForgotPassword from './pages/forgot-password'
 import Login from './pages/login'
